adding compensation unit test

+++ src/main/java/com/mindex/challenge/data/Compensation.java
@@ -4,7 +4,7 @@ import java.util.Date;
 
 public class Compensation {
 	private Employee employee;
-	private float salary;
+	private Double salary;
 	private Date effectiveDate;
 	
 	public Compensation() {		
@@ -18,11 +18,11 @@ public class Compensation {
 		this.employee = employee;
 	}
 	
-	public float getSalary() {
+	public Double getSalary() {
 		return salary;		
 	}
 	
-	public void setSalary(float salary) {
+	public void setSalary(Double salary) {
 		this.salary = salary;
 	}
